+++ change_log.md
@@ -1,6 +1,13 @@
 ### Test project for vertical grid emulation in VTV
 
 --------
+0.0.14
+- fix error for *.dat file path for linux
+- implemented column width resize by programmaticaly
+
+![](img/resize_column.gif)
+
+--------
 0.0.13
 - changed header height, ~~font style and~~ font color